Fix compiler warning messages
Various occurances of the following compiler warnings have been
fixed:
* Type mismatch
* Misleading code indentation
* Array bounds violation warning in blastest when using gcc 11
  without -fPIC flag

AMD-Internal: [CPUPL-5895]

+++ addon/aocl_gemm/aocl_gemm_check.h
@@ -169,7 +169,7 @@
     { \
         char print_msg[ 150 ]; \
  \
-        sprintf( print_msg, "** On entry to %6s, parameter number %2i of problem %ld had an illegal value", op_str, info, gemm_no); \
+        sprintf( print_msg, "** On entry to %6s, parameter number %2i of problem %ld had an illegal value", op_str, info, (long int) gemm_no); \
         bli_print_msg(print_msg, __FILE__, __LINE__); \
         err_no = info; \
     } \
